chore: directory structure (#104)
