+++ src/lib.rs
@@ -4,14 +4,12 @@
 #![allow(clippy::too_many_arguments)]
 #![allow(clippy::type_complexity)]
 
-mod algorithm;
 mod datatype;
-mod gucs;
-mod index;
 mod postgres;
+mod projection;
 mod types;
 mod upgrade;
-mod utils;
+mod vchordrq;
 
 pgrx::pg_module_magic!();
 pgrx::extension_sql_file!("./sql/bootstrap.sql", bootstrap);
@@ -24,8 +22,12 @@ unsafe extern "C" fn _PG_init() {
     }
     detect::init();
     unsafe {
-        index::init();
-        gucs::init();
+        vchordrq::init();
+
+        #[cfg(any(feature = "pg13", feature = "pg14"))]
+        pgrx::pg_sys::EmitWarningsOnPlaceholders(c"vchord".as_ptr());
+        #[cfg(any(feature = "pg15", feature = "pg16", feature = "pg17"))]
+        pgrx::pg_sys::MarkGUCPrefixReserved(c"vchord".as_ptr());
     }
 }
