+++ src/vchordrq/algorithm/k_means.rs
@@ -1,4 +1,4 @@
-use crate::algorithm::parallelism::{ParallelIterator, Parallelism};
+use crate::vchordrq::algorithm::parallelism::{ParallelIterator, Parallelism};
 use base::scalar::*;
 use half::f16;
 use rand::rngs::StdRng;
