+++ src/vchordrq/algorithm/tuples.rs
