+++ src/vchordrq/mod.rs
@@ -1,0 +1,10 @@
+mod algorithm;
+mod gucs;
+mod index;
+
+pub unsafe fn init() {
+    unsafe {
+        index::init();
+        gucs::init();
+    }
+}
